Use HTMLBars compiler directly.
Removes large issues with caching and shared plugin arrays that occurs
in Ember's template compiler.  We do not need the output templates to be
compatible (the AST is perfectly fine for our needs). This makes it
possible to use this as a stand alone package if we want to extract...