Make the block-indentation rule aware of leading content
If the block children are inline elements, then there can be important reasons for them to appear on the same line. For example:

```javascript
<div>
  <span class="key">key</span>:<span class="value">value</span>
</div>
```

will render as

```
key: value
```

but

```javascript
<div>
  <span class="key">key</span>:
  <span class="value">value</span>
</div>
```

will render as

```
key:
value
```

The block indentation rule was previously forbidding the first version because the second ```<span>```'s column number is not two. This commit makes the indentation rule aware of previous content on the same line, whether it is an element/mustache node or text.