Remove some roles from interactive roles listing.
Even those these are listed as "widgets" in the list
[here](https://www.w3.org/TR/wai-aria/complete#widget_roles) their
descriptions do not discuss interactivity, and they do not extend from
an interactive superclass role.

It is possible that some of these will need to be brought back with more
nuance (i.e. they may be considered interactive if other attributes are
also present).