Make unit ember-template-lint-tes run on windows
Previously the unit tests that ran ember-template-lint using node failed on windows, due to using forward slashes instead of backslashes as path seperator. But also it is not possible to just execute a .js file in windows. You have to provide a "real" executable to run, in this case node.

So I changed the uses of exec to use execFile with node as the first argument. This was not strictly needed, but execFile has better support for passing arguments and is a bit more efficient, so I saw it as an opputunity to change that aswell.