Ensure contextual components do not break block indentation.
When the template linter processes each file it must convert contextual
component invocation into `{{#component` to avoid errors in glimmer-vm
when parsing. Unfortunately, when this happens our calculations for
where the end of the block should be were thrown off (because
`component` is longer/shorter than the actually invoked component).