Add rw module
Node ability to synchronously read from stdin is problematic to say the
least. Some constraints on how to do this are that we still support Node
version 6 and this needs to run on Windows.

The code base is very much not ready for async and to do all callbacks
would be a refactoring nightmare. This lead me to look into the rw
module which is very popular on NPM and written by the author of d3.

Looking though the code shows that it is compatible with the node
versions we use (no deprecations warnings) and it handles all the
caveats for reading on stdin.

The only warning is that ember-template-lint cannot use process.stdin on
its own anymore. But considering the use case I don't see that need.