Add "octane" configuration preset.
This config represents what we believe will be the "recommended" config
for the next major version of ember-template-lint which will target
Ember's Octane Edition.

At the moment we are adding the configuration to a specific `octane`
preset which will allow users to begin using it immediately, but once
Octane "ships" we will release a major version bump of
ember-template-lint changing `recommended` to use this new config.

Co-authored-by: Melanie Sumner <melaniersumner@gmail.com>