Refactor argv processing in main bin script
Prior to this change, we had a very simplistic argument processing
pattern. By adding another argument it was time to upgrade. However, we
did not wish to bring in a full parsing library but instead wanted to
have a simple parsing function that new how to pull the argv into
a named and positional object look up (perfect for easy destructuring).

I use a simplified [getopt][1] like pattern when writing shell (Bash)
scripts and felt the pattern was also useful in JavaScript (which
I think it was).

I also noticed that we are now using ES6 features so I took advantage of
that in my refactoring. JavaScript is so cool.

[1]: https://tritarget.org/#getopt.sh