Fix issues introduced by ember-template-recast@4.0.1.
The fixes introduced into the `next` branch (while merging master into
the branch) which uses ember-template-recast as its parser). The
underlying reason for the issue is that ember-template-lint was
_basically_ relying on the bug fixed (e.g. it was looking at
`AttrNode`'s `.value` and doing `sourceForLoc` on that node, then
checking if it starts with a quote).

In order for these rules to function (`quotes` and
`no-quoteless-attribute`) they need to know if a given attribute had a
value (aka was it `<div data-test-blah></div>` vs `<div
data-test-blah="thing"></div>`) and what type of quotes were used
(single, double, none, etc).

This updates ember-template-recast to 4.1.0 (which adds some additional
properties to `AttrNode`'s) and tweaks the implementation of the rules
to leverage that functionality.

Additionally, I updated the `quotes` rule to emit a bit more "source"
information (which I found a bit more helpful to have more context).