docs: add script to generate rules table in README
Before this change, it was difficult to find the list of rules in this package, and it was also hard to find what config/preset each rule was a part of. Those are two of the most important pieces of information when visiting a linting plugin's homepage.

After this change, there is a new autogenerated table in README.md with the complete list of rules and emojis to indicate what config/preset each rule is part of.

`yarn update` will automatically update the table (this script is based on a script from eslint-plugin-ember). I have also added a CI test to ensure this script is always run when changes are made.