Use a copied version of plugins.configurations
When creating a new instance of Linter, `lib/get-config`'s
processPlugins functions is called.

But processPlugins will modify config.plugins.configurations object.

Since the same plugins.configurations object is used from one test to
another in the plugin test harness, a test config would be wrong when
passed to the next test.

To isolate a plugin's test one from the other, this commit makes a deep
copy of plugins.configurations.