map tagnames to their equivalent roles
If we're seeing if a <div> has the equivalent role to a "header",
for example, we need to look for role="banner", not role="header".

I added a lookup table between tag names and their equivalent aria
roles.

Now, in order to find the equivalent landmark type we have the
following hierarchy:

1) Find the explicit role attribute on the HTML node
2) Look up the equivalent role in a lookup table
3) If no match is found, return the tag name of the HTML element