Ensure globbing is only used when using a globlike pattern.
Within the command line script's `expandFileGlobs` method we were using
`is-valid-glob` to determine if a given possible pattern / filename
string was in fact a pattern or not. The idea here is that if a possible
pattern does not contain any glob related wildcards or matchers **and**
it is a literal file (ending with `.hbs`) it should be used directly
and we should avoid globbing.

The most common scenario for this is when using built-in editor tooling
to lint _just_ the current file. In that scenario we **never** want to
waste the overhead of globbing, since we were given an explicit path.

Unfortunately, the previous code used `is-valid-glob` and therefore
**always** considered any pattern as "valid" since `isValidGlob` is
_basically_ just:

```js
function isValidGlob(pattern) {
  return Boolean(pattern) && (Array.isArray(pattern) || typeof pattern === 'string');
}
```

The package we _should have used_ is
[is-glob](https://www.npmjs.com/package/is-glob), which is actually
checking the contents of the string to see if it contains globlike
matchers.