Track duplicate ID's within conditionals.
Introduce a custom conditional scope tracking system to track ids used
within conditional branches. This is similar to the scope tracking that
we do for identifying unused locals, but with some additonal
"complications".

The goal is to ensure that ids used within a conditional can be used in
_either_ branch (not both) **and** cannot be used _after_ the
conditional statement (e.g. in the root of the template after a
conditional).

Co-authored-by: Robert Jackson <me@rwjblue.com>
Co-authored-by: Melanie Sumner <melaniersumner@gmail.com>