Add `no-dynamic-subexpression-invocations`.
When using Ember versions prior to 3.25 the usage of dynamic invocations
for helpers and modifiers did not work. Unfortunately, some versions of
Ember silently ignored additional arguments (3.16) wherease others throw
a very bizarre error (3.20) when you attempt to invoke what the
rendering engine knows as a dynamic value as if it were a helper.

For example, in versions of Ember prior to 3.25 the helper invocation
here is impossible:

```hbs
{{#if (this.someHelper)}}
  Hi!
{{/if}}
```

This rule helps applications using Ember versions prior to 3.25 avoid
these types of invocation.