Remove `deprecated-each-syntax` from `recommended` config
This rule was written to help folks migrate away from these older
`{{#each` syntaxes in the 1.x era. The assumption here is that if you
have `{{#each something}}` that you are relying on "context shifting",
but that is simply not the case. For example, you may use a `repeat`
helper like:

```hbs
{{#each (repeat 3)}}
 do this three times
{{/each}}
```

In that case, you may not _need_ the block param (which would make this
not trigger the rule, but would also trigger the
"no-unused-block-params" rule).

This rule should be deprecated itself, and slated for removal in
ember-template-lint@4.