Emit full path when emitting output on CI.
When running in CI environments, it is helpful to emit the full path to
the file on disk (instead of the path relative to the specified working
directory). For example, in GitHub Actions CI there are a number of
matchers (RegExp's on standard out) that they use to add annotations to
files.

Prior to this change, those annotations worked properly for applications
that were located in the repo root path but failed to properly identify
a file when the the application was not in the repo root. Specifically,
when linting in a monorepo format (basically when the project's
`package.json` is not also the location of `.git`) the annotations would
not work properly.

This changes the output to always use the full path in CI conditions.
This should not regress users (seeing the relative path _looks_ better)
since they don't often run in CI environments.