no-curly-component-invocation: add `unique-id` to allowed-list
In Ember v4.4.0, a `unique-id` helper was added which generates a unique
ID string suitable for use as an ID attribute in the DOM. When this
helper is used as:

`<p>{{unique-id}}</p>`

the no-curly-component-invocation rule will flag it with the message:

```
You are using the component {{unique-id}} with curly component syntax. You
should use <UniqueId> instead. If it is actually a helper you must manually add
it to the 'no-curly-component-invocation' rule configuration, e.g.
`'no-curly-component-invocation': { allow: ['unique-id'] }`.
```

As this is a perfectly legal usage of the built-in helper, let's go
ahead and add it to the rule's allowed-list such that it no longer
raises it as an issue.