no-implicit-this: add `unique-id` helper to allowed-list
In Ember v4.4.0, a `unique-id` helper was added which generates a unique
ID string suitable for use as an ID attribute in the DOM. When this
helper is used as:

`<p>{{unique-id}}</p>`

the no-implicit-this rule will flag it with the
message:

```
Ambiguous path 'unique-id' is not allowed. Use '@unique-id' if it is a named
argument or 'this.unique-id' if it is a property on 'this'. If it is a helper
or component that has no arguments, you must either convert it to an angle
bracket invocation or manually add it to the 'no-implicit-this' rule
configuration, e.g. 'no-implicit-this': { allow: ['unique-id'] }.
```

As this is a perfectly legal usage of the built-in helper, let's go
ahead and add it to the rule's allowed-list such that it no longer
raises it as an issue.