Remove usage of `cross-env`
We were using `cross-env` in order to set the
`--experimental-vm-modules` flag via `process.env.NODE_OPTIONS` to run
`jest`. Unfortunately, this means that the `NODE_OPTIONS` that was being
set in our `.github/workflows/ci.yaml` (which increases
`--max-old-space-size` to 4096) is completely clobbered.

This doesn't **fix** the memory issues that we are seeing with Node 18,
but it does bring back a previously known good work around so that CI is
"green again".

Co-authored-by: Steve Calvert <steve.calvert@gmail.com>