link-to: allow only model or models
The no-unknown-arguments-for-builtin-components lint rule is triggered
when a <LinkTo /> component only has the `@model` or `@models` argument,
stating that it must have either `@route` or `@query` as well.

The <LinkTo /> component supports passing only the `@model` or `@models`
argumment. In this case, it will simply reuse the current route /
queryParams.

As such, we should update the linter rule to allow this case.

This closes https://github.com/ember-template-lint/ember-template-lint/issues/2700.