Link to "skipping helpers" part of the codemod doc in `no-curly-component-invocation` docs (#2732)
In order to ignore custom helpers in the `no-curly-component-invocation`
linter rule, a developer manually must add these to the config `allow`
list. The `ember-angle-brackets-codemod` README has a useful snippet
that can be run in the console in order to generate a list of helpers in
a given ember app.

Let's add a reference to that here in case folks want to use it to
configure this lint rule too.

This closes https://github.com/ember-template-lint/ember-template-lint/issues/2286.