fixing wrong tests
I'm pretty sure these tests were incorrect and it was just hidden by another bug.

The checkbox role *does* support aria-invalid. So I changed that one to button role, which does not.

Also, the menuitem element (not to be confused with the menuitem role) does not appear at all in the set of elementRoles from aria-query. So I changed that test to use `<a role="menuitem">` instead. The menuitem element is [deprecated and non-standard](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/menuitem).