fix: Don't require redundant role for <img alt="" />
Fixes https://github.com/ember-template-lint/ember-template-lint/issues/2951

This updates the `require-valid-alt-text` rule so that it does not
*require* setting `role="presentation"` or `role="none"` for an `<img
alt="" />`.

Context

After bumping `aria-query` in https://github.com/ember-template-lint/ember-template-lint/pull/2927,
`ember-template-lint` began reporting an error for the following:

```html
<img alt="" role="presentation" />
```

The error is from the `no-redundant-role`. I believe it is correct to
report an error in this case because `<img alt="" />` already implicitly
has a `role` of `presentation`.

However, this new behavior was in direct conflict with part of the
`require-valid-alt-text` rule that required setting
`role="presentation"` or `role="none"` whenever an image has `alt=""`.