Fix `--no-error-on-unmatched-pattern` for multiple patterns
Previously, when the cli was passed multiple patterns and `--no-error-on-unmatched-pattern` was set, any failing pattern would cause the entire CLI call to be aborted. This commit updates things so that the logic is applied on a per-pattern basis, so that any matching patterns are still checked.

This brings ember-template-lint into line with ESLint's `--no-error-on-unmatched-pattern` flag.

Resolves #2758